src/requirements.txt
click==8.1.3
Flask==2.1.2
Flask-JWT-Extended==4.4.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
PyJWT==2.3.0
pyparsing==3.0.8
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0